Update Legacy Application Flow documentation
I added a basic description of the Legacy Application Flow. It's pretty simple, but it took a bit of tooling around so this should save some people the trouble.

I discovered that client_id is not passed from the OAuth2Session or LegacyApplicationClient when fetch_token is called, which I believe to be a bug.