[bugfix] Add headers in refresh_token flow
Indeed, we are sending default headers when fetching a token. I thought
it was logic to also send these same default headers (and offer our
users the possibility to set their own headers) in the `refresh_token`
method.

Also, one of the main reason for this is that some applications (the
Flowdock API for instance) do need those headers to return the
appropriate content :)