Register a new compliance hook for token_request
I'm working with an oauth2 server we know is broken, however it works
for other systems and the company has no interest in fixing it. It
requires that I submit auth in JSON format rather than form encoded. We
shouldn't support this specific problem, but if we have a compliance
hook that allows me to change the sending format of the payload as
required I can fix my problem and potentially others in future.

Related to: #244 (but not the reason for the patch)